Extract common text utils
