+++ src/app/search/plug-search.ts
@@ -1,7 +1,7 @@
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { DimLanguage } from 'app/i18n';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
-import { startWordRegexp } from './search-filters/freeform';
+import { startWordRegexp } from './text-utils';
 
 export function createPlugSearchPredicate(
   query: string,
